pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = ""
authors = ["Pablo Thasi <thasipablo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
